Add TrackPath option, output absolute/relative path also upon impact
Track the relative movement from a particle starting position to the current position or impact. Re-use the TrackDispersion arrays which tracks the absolute particle path, thereby making TrackPath and TrackDispersion mutually exclusive. Expection is that there a little to no cases where both options are required at the same time.