Calculate memory usage using Unique Set Size (USS)
The Unique Set Size, or USS, represents the private memory of a process. In other words, it shows libraries and pages allocated only to this process.

This type of memory is also very useful, as it represents the actual cost of launching another process. And, vice versa, when a process ends, it represents the actual amount that is returned to the system.